Moving each component into its own directory.
