--- src/components/Profile.js
@@ -1,8 +1,0 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-
-class profile extends Component {
-
-}
-
-export default connect()(profile);
