+++ src/components/PromotionItem/PromotionItem.js
